// packages/meteor-common/src/modules_app_core/state/utils/state_tasks.enums.ts
export enum ECommonStateTaskErrorEndTags {
  malformed_account_id = "malformed_account_id",
  incorrect_password = "incorrect_password",
  unknown_error = "unknown_error",
  account_not_found = "account_not_found",
  account_key_unknown_to_meteor = "account_key_unknown_to_meteor",
  account_key_type_mismatch = "account_key_type_mismatch",
  account_key_already_exists = "account_key_already_exists",
  account_key_is_not_seed_phrase = "account_key_is_not_seed_phrase",
  account_key_doesnt_match_public_key = "account_key_doesnt_match_public_key",
}